+++ lib/lazy_memo.dart
@@ -3,3 +3,4 @@
 
 export 'src/lazy.dart';
 export 'src/memoized_functions.dart';
+export 'src/utils.dart';