+++ tool/actions.sh
@@ -56,4 +56,4 @@ dart example/bin/lazy_example.dart
 echo
 echo -e "${GREEN}=== Running Benchmarks $PWD/benchmark ${RESET}"
 echo
-dart benchmark/bin/lazy_memo_benchmark.dart
+dart benchmark/bin/lazy_benchmark.dart